requirements.txt
# Django et REST framework
django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
gunicorn==20.1.0
whitenoise==6.4.0

# Traitement d'images
pillow==9.5.0
opencv-python==4.7.0.72

# Machine Learning
tensorflow==2.12.0
numpy==1.23.5
pandas==2.0.1
scikit-learn==1.2.2
scipy==1.10.1
SimpleITK==2.2.1

# Visualisation
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
dash==2.9.3
dash-bootstrap-components==1.4.1
ipympl==0.9.3
ipywidgets==8.0.6

# Développement
jupyter==1.0.0

# Utilitaires
tqdm==4.65.0
python-dateutil==2.8.2